rewrite by chatgpt
